トゲ、トゲの台、拡散弾発射装置をマップ上に表示させるようにした(トゲとトゲの台はコメント化してます)

+++ ACG/ACG/ObjNeedleStand.cpp
@@ -9,10 +9,10 @@
 using namespace GameL;
 
 //コンストラクタ
-CObjNeedleStand::CObjNeedleStand(float x, float y)
+CObjNeedleStand::CObjNeedleStand(int x, int y)
 {
-	m_px = x * BLOCK_SIZE;
-	m_py = y * BLOCK_SIZE;
+	m_px = (float)x * BLOCK_SIZE;
+	m_py = (float)y * BLOCK_SIZE;
 }
 
 //イニシャライズ
